avoid compiler warning (implicit cast from "int" to "char")
SVN-Revision: 27224

+++ contrib/brl/bbas/bdgl/tests/test_region_algs.cxx
@@ -14,7 +14,7 @@ static void test_region_algs()
   unsigned short * I1 = new unsigned short[n1];
   for (unsigned int i = 0; i<n1; i++)
   {
-    X1[i] = 1.0f*i; Y1[i] = 2.0f*i; I1[i]= 10*i;
+    X1[i] = 1.0f*i; Y1[i] = 2.0f*i; I1[i]= (unsigned short)(10*i);
   }
 
   vdgl_digital_region_sptr r1 = new vdgl_digital_region(n1, X1, Y1, I1);
@@ -27,7 +27,7 @@ static void test_region_algs()
   unsigned short * I2 = new unsigned short[n2];
   for (unsigned int i = 0; i<n2; i++)
   {
-    X2[i] = 3.0f*i; Y2[i] = -6.0f*i; I2[i]= 100*i;
+    X2[i] = 3.0f*i; Y2[i] = -6.0f*i; I2[i]= (unsigned short)(100*i);
   }
   vdgl_digital_region_sptr r2 = new vdgl_digital_region(n2, X2, Y2, I2);
   delete[] X2;
@@ -56,7 +56,6 @@ static void test_region_algs()
     TEST("bdgl_region_algs::merge(): Y value", rm->Y(), r2->Y());
     TEST("bdgl_region_algs::merge(): I value", rm->I(), r2->I());
   }
-  //TEST("Num Intersection points = 1", pts.size(), 1);
 }
 
 TESTMAIN(test_region_algs);
